+++ Lib/SimulationCodes/Common/__init__.py
@@ -1,3 +1,4 @@
 """Common objects to iHIBPsim and the new INPASIM and FILDSIM codes."""
 from Lib.SimulationCodes.Common.fields import fields as Fields
 from Lib.SimulationCodes.Common.geometry import Geometry
+from Lib.SimulationCodes.Common.strikes import Strikes