+++ Lib/SimulationCodes/__init__.py
@@ -1,0 +1,4 @@
+"""
+Contains routines to interact with the simulation codes
+"""
+