+++ Lib/__init__.py
@@ -175,5 +175,3 @@ try:
 except:
     logger.warning('28: It was not possible to initialise the plotting ' +
                    'settings')
-
-