+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.2.0: Melon con Jamon
+## General
+- Changed @ for : in documentations
+## Data
+- INPA object documentation improved
+- getGeomShots() removed from INPA object
+## Simulation codes
+- Added compatibility with SINPA 4
+
+
 # 1.1.0: Melon con Jamon
 ## FIDASIM library
 - FIDASIM library distributed with the suite is no longer mantained. Please download it from the main repo of FIDASIM4